Split up raytracing modules
