+++ src/command_buffer/traits.rs
@@ -8,7 +8,7 @@ use crate::command_buffer::CommandBuffer;
 use crate::core::queue::Queue;
 use crate::domain::ExecutionDomain;
 use crate::query_pool::{AccelerationStructurePropertyQuery, QueryPool};
-use crate::raytracing::acceleration_structure::{AccelerationStructure, AccelerationStructureBuildInfo};
+use crate::raytracing::*;
 
 /// Trait representing a command buffer that supports transfer commands.
 pub trait TransferCmdBuffer {
