+++ src/prelude.rs
@@ -33,13 +33,17 @@ pub use crate::pipeline::shader::ShaderCreateInfo;
 pub use crate::resource::*;
 pub use crate::resource::buffer::{Buffer, BufferView};
 pub use crate::resource::image::{Image, ImageView};
+pub use crate::resource::query_pool::*;
+pub use crate::resource::raytracing::*;
 pub use crate::sampler::Sampler;
 pub use crate::sync::execution_manager::ExecutionManager;
 pub use crate::sync::fence::*;
 pub use crate::sync::semaphore::*;
 pub use crate::sync::thread_context::ThreadContext;
+pub use crate::util::address::*;
 pub use crate::util::byte_size::ByteSize;
 pub use crate::util::deferred_delete::DeletionQueue;
+pub use crate::util::transform::TransformMatrix;
 pub use crate::util::upload::staged_buffer_upload;
 pub use crate::wsi::frame::{FrameManager, InFlightContext};
 pub use crate::wsi::surface::Surface;
